User progress reset buttons function as desired
User must press the desired reset button, then press the Apply button at the bottom to confirm their choice. This is to help prevent accidental loss of progress by mistake. The button highlighting is a bit derpy, but it is just a visual bug. I can fix it in the future.